fix(consumer): Pass batch time as correct units to streaming strategy
This also allows expressing the max_batch_time as a float instead of an
integer to represent subsecond values

+++ snuba/consumer.py
@@ -278,7 +278,7 @@ class StreamingConsumerStrategyFactory(ProcessingStrategyFactory[KafkaPayload]):
         processor: MessageProcessor,
         writer: BatchWriter[WriterTableRow],
         max_batch_size: int,
-        max_batch_time: int,
+        max_batch_time: float,
         replacements_producer: Optional[ConfluentKafkaProducer] = None,
         replacements_topic: Optional[Topic] = None,
     ) -> None:
